Extract generic web preview refresh dispatch
